requirements.txt
seaborn==0.13.2
torchvision==0.20.1
matplotlib==3.8.0
numpy==1.23.4
torch==2.5.1
tqdm==4.67.1
patchify==0.2.3
albumentations==1.3.0
pandas==2.1.1
segmentation_models_pytorch==0.3.4
scikit-learn==1.1.3
scikit-image==0.19.3
opencv-python==4.7.0.68
torch-summary==1.4.5
pillow==9.3.0